service/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2025.11.12
charset-normalizer==3.4.4
DateTime==6.0
distro==1.9.0
groq==0.36.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
outcome==1.3.0.post0
pydantic==2.12.5
pydantic_core==2.41.5
PySocks==1.7.1
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
selenium==4.38.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.9.0
wsproto==1.3.2
zope.interface==8.1.1
